Extracted ShortcutsMap class
