+++ src/main/java/address/shortcuts/GlobalHotkey.java
@@ -6,11 +6,19 @@ import address.events.BaseEvent;
  * Represents a global hotkey
  */
 public class GlobalHotkey {
-    String hotkeyString;
-    BaseEvent eventToRaise;
+    private String hotkeyString;
+    private BaseEvent eventToRaise;
 
     GlobalHotkey(String hotkeyString, BaseEvent eventToRaise){
         this.hotkeyString = hotkeyString;
         this.eventToRaise = eventToRaise;
+    }
+
+    protected String getHotkeyString() {
+        return hotkeyString;
+    }
+
+    protected BaseEvent getEventToRaise() {
+        return eventToRaise;
     }
 }
